Users details page. Twitter bootstrap.

+++ app/models/Role.scala
@@ -7,7 +7,7 @@ import play.api.libs.json._
  * ndidialaneme
  */
 
-case class Role(id: Int) extends Model[Role]
+case class Role(id: Long) extends Model[Role]
 
 object Role{
   implicit object RoleFormat extends Format[Role] {
